+++ packages/sdk/test/client/resources/expectedJsons.ts
@@ -543,6 +543,7 @@ export const seqNums: NextUpdateSequenceNumbers = {
     blockEnergyLimit: 1n,
     minBlockTime: 1n,
     finalizationCommiteeParameters: 1n,
+    validatorScoreParameters: 1n,
 };
 
 export const specialEventList: BlockSpecialEvent[] = [
@@ -1405,6 +1406,7 @@ export const bakerAccountInfo: AccountInfoBaker = {
         bakerSignatureVerifyKey: 'c385ccb5c8a0710a162f2c107123744650ff35f00040bfa262d974bfb3c3f8f1',
         restakeEarnings: true,
         stakedAmount: CcdAmount.fromMicroCcd(7349646704751788n),
+        isSuspended: false,
     },
     accountCooldowns: [],
     accountAvailableBalance: CcdAmount.fromMicroCcd(100000000000000n),
@@ -1922,8 +1924,11 @@ export const bakerPoolStatus: BakerPoolStatus = {
             finalizationCommission: 1,
             transactionCommission: 0.1,
         },
+        missedRounds: 0n,
+        isPrimedForSuspension: false,
     },
     allPoolTotalCapital: CcdAmount.fromMicroCcd(46071942529284135n),
+    isSuspended: false,
 };
 
 export const passiveDelegationStatus: PassiveDelegationStatus = {
@@ -1960,6 +1965,7 @@ export const bakerPoolStatusWithPendingChange: BakerPoolStatus = {
         effectiveTime: new Date('2022-12-08T07:54:00.000Z'),
     },
     allPoolTotalCapital: CcdAmount.fromMicroCcd(46470271917743628n),
+    isSuspended: false,
 };
 
 export const invokeContractResult: InvokeContractResult = {